% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial3level.h.R
\name{factorial3level}
\alias{factorial3level}
\title{Factorial Design - 3 Levels}
\usage{
factorial3level(
  data,
  factorA,
  factorB,
  factorC,
  dep,
  varEq = TRUE,
  showPlot = TRUE,
  showANOVA = TRUE
)
}
\arguments{
\item{data}{.}

\item{factorA}{.}

\item{factorB}{.}

\item{factorC}{.}

\item{dep}{.}

\item{varEq}{.}

\item{showPlot}{.}

\item{showANOVA}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$anovaTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$effects} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$anovaTable$asDF}

\code{as.data.frame(results$anovaTable)}
}
\description{
Factorial Design - 3 Levels
}
